Revert "push inverse det and solve into Decomposition"
This reverts commit 8f078ec1597a08cfa6e2f7a76ac671ce1f310466.

Conflicts:

	src/matrix/mod.rs
